# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Liste des users",
                "description": "Returns list of users",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserListResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "operationId": "8c72d9a9a5dce4c3e4d320a471e80c58",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de la resource",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserListResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserListResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{user}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Met à jour un utilisateur existant",
                "description": "Met à jour les informations d'un utilisateur ainsi que ses rôles et son avatar.",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/schemas/UserListResource"
                },
                "responses": {
                    "202": {
                        "description": "Utilisateur mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserListResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies sont invalides."
                                        },
                                        "errors": {
                                            "description": "Détails des erreurs de validation",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Supprime un utilisateur",
                "description": "Supprime un utilisateur existant par son ID.",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Utilisateur supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "No Content"
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vous n'avez pas les permissions nécessaires pour effectuer cette action."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserListResource": {
                "title": "UserListResource",
                "description": "User resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "ID",
                                    "type": "integer"
                                },
                                "firstname": {
                                    "description": "First Name",
                                    "type": "string"
                                },
                                "lastname": {
                                    "description": "Last Name",
                                    "type": "string"
                                },
                                "fullname": {
                                    "description": "Full Name",
                                    "type": "string"
                                },
                                "email": {
                                    "description": "Email",
                                    "type": "string",
                                    "format": "email"
                                },
                                "phone_number": {
                                    "description": "Phone Number",
                                    "type": "string"
                                },
                                "roles": {
                                    "description": "Roles",
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "avatar": {
                                    "description": "Avatar URL",
                                    "type": "string"
                                },
                                "created_at": {
                                    "description": "Created At",
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserListResource"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}